k-25-05-2022/basic_folder_d/books/urls.py
from django.urls import path

from .views import listview, details

app_name = "books"
urlpatterns = [
    path("books/", listview, name ="list"),
    path("books/<int:book_id>", details, name="details"),
]