k-25-05-2022/basic_folder_d/dev-requirements.txt
#dev
ipython
django-extensions

-r requirements.txt